perf(poker): lazy-load the table window, seed deck from rng
The gate in features/PokerTable/index.tsx now React.lazy-imports
PokerTableWindow (engine, bots, pokersolver, components) as its own
webpack chunk, so nothing ships in the main bundle until the table is
opened. The controller also passes its rng to the engine's deck, which
makes seeded test runs deterministic (the run-out test was flaky).

Pre-commit hook ran. Total eslint: 0, total circular: 0
